Create TestBoundsParser, split ParameterFactory
Breaking change: previously, {"inputMap": map, "bounded", bounded}
would be written to the JSON file. Now, only the original map is written.

Code for parsing test bounds is now in TestBoundsParser. Other code still
needs to be refactored so that all unchecked casts are moved there, and
all raw types are removed.

ParameterFactory is now an abstract class, implemented by
DependentParameterFactory and FixedParameterFactory. This is to avoid
having a class with completely different behavior based on the type being
used. The correct class is chosen by TestBoundsParser